w3c-reports/12092014/mac-wk2/runned-tests/webmessaging/postMessage_solidus_xorigin.sub-expected.txt

CONSOLE MESSAGE: Unable to post message to http://localhost:8800. Recipient has origin http://www1.localhost:8800.


Harness Error. harness_status.status = 2 , harness_status.message = null
